backend/requirements.txt
scikit-learn>=1.0.0
numpy>=1.20.0
apiflask>=2.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
pymysql>=1.1.0
cryptography>=41.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
